refactor: rename custom hook filename
